# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_exnodes_hrm-api_internal_dto.EmergencyContactInput:
    properties:
      full_name:
        maxLength: 200
        minLength: 1
        type: string
      phone_number:
        maxLength: 50
        type: string
      relationship:
        maxLength: 100
        type: string
    required:
    - full_name
    type: object
  github_com_exnodes_hrm-api_internal_dto.EmployeeCreate:
    properties:
      bank_account:
        type: string
      bank_holder_name:
        type: string
      bank_name:
        type: string
      basic_salary:
        minimum: 0
        type: number
      contract_end_date:
        type: string
      contract_renewal:
        type: boolean
      contract_sign_date:
        type: string
      contract_type:
        description: Contract / salary / bank
        type: string
      current_address:
        type: string
      cv_url:
        type: string
      department_id:
        description: Work
        type: string
      dob:
        type: string
      education:
        enum:
        - high_school
        - college
        - bachelor
        - master
        - doctorate
        type: string
      email:
        description: User credentials (created in tx)
        type: string
      emergency_contacts:
        description: 'Emergency contacts — full replacement list at create (employees
          parity #4).'
        items:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.EmergencyContactInput'
        type: array
      experience_year:
        minimum: 0
        type: integer
      full_name:
        description: HR personal info
        maxLength: 200
        minLength: 1
        type: string
      gender:
        enum:
        - male
        - female
        - other
        type: string
      id_back_image:
        type: string
      id_front_image:
        type: string
      id_issue_date:
        type: string
      id_number:
        type: string
      insurance_salary:
        minimum: 0
        type: number
      is_active:
        type: boolean
      join_date:
        type: string
      manager_id:
        type: string
      marital_status:
        enum:
        - single
        - married
        - other
        type: string
      nationality:
        type: string
      password:
        minLength: 8
        type: string
      payment_method:
        type: string
      permanent_address:
        type: string
      personal_email:
        type: string
      phone:
        type: string
      position_id:
        type: string
      role_ids:
        description: Roles assigned at creation
        items:
          type: string
        type: array
    required:
    - email
    - full_name
    - password
    type: object
  github_com_exnodes_hrm-api_internal_dto.EmployeeSelfUpdate:
    properties:
      current_address:
        type: string
      dob:
        type: string
      emergency_contacts:
        description: |-
          Emergency contacts — self may manage their own list (pointer-to-slice:
          nil = leave unchanged, [] = clear all, non-empty = replace the set).
        items:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.EmergencyContactInput'
        type: array
      full_name:
        maxLength: 200
        minLength: 1
        type: string
      gender:
        enum:
        - male
        - female
        - other
        type: string
      marital_status:
        enum:
        - single
        - married
        - other
        type: string
      permanent_address:
        type: string
      personal_email:
        type: string
      phone:
        type: string
    type: object
...
  github_com_exnodes_hrm-api_internal_dto.EmployeeUpdate:
    properties:
      bank_account:
        type: string
      bank_holder_name:
        type: string
      bank_name:
        type: string
      basic_salary:
        minimum: 0
        type: number
      clear_department:
        type: boolean
      clear_manager:
        type: boolean
      clear_position:
        type: boolean
      contract_end_date:
        type: string
      contract_renewal:
        type: boolean
      contract_sign_date:
        type: string
      contract_type:
        type: string
      current_address:
        type: string
      cv_url:
        type: string
      department_id:
        type: string
      dob:
        type: string
      education:
        enum:
        - high_school
        - college
        - bachelor
        - master
        - doctorate
        type: string
      emergency_contacts:
        description: |-
          Emergency contacts — pointer-to-slice partial-PATCH semantics:
          nil/absent = leave unchanged, [] = clear all, non-empty = replace the set.
        items:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.EmergencyContactInput'
        type: array
      experience_year:
        minimum: 0
        type: integer
      full_name:
        type: string
      gender:
        enum:
        - male
        - female
        - other
        type: string
      id_back_image:
        type: string
      id_front_image:
        type: string
      id_issue_date:
        type: string
      id_number:
        type: string
      insurance_salary:
        minimum: 0
        type: number
      is_active:
        description: toggles user.is_active
        type: boolean
      join_date:
        type: string
      manager_id:
        type: string
      marital_status:
        enum:
        - single
        - married
        - other
        type: string
      nationality:
        type: string
      payment_method:
        type: string
      permanent_address:
        type: string
      personal_email:
        type: string
      phone:
        type: string
      position_id:
        type: string
    type: object
...
  github_com_exnodes_hrm-api_internal_dto.LeaveQuotaUpdateRequest:
    properties:
      annual_leave_quota:
        maximum: 365
        minimum: 0
        type: number
      sick_leave_quota:
        maximum: 365
        minimum: 0
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: HRM API for Exnodes — see /docs/superpowers/specs/2026-05-15-go-migration-design.md
    for the migration plan.
  title: Exnodes HRM API
  version: 0.1.0
paths:
...
  /api/v1/employees:
    get:
      parameters:
      - description: page (default 1)
        in: query
        name: page
        type: integer
      - description: page size (default 20, max 100)
        in: query
        name: page_size
        type: integer
      - description: free text (full_name/phone/personal_email/user.email)
        in: query
        name: search
        type: string
      - description: department uuid
        in: query
        name: department_id
        type: string
      - description: position uuid
        in: query
        name: position_id
        type: string
      - description: manager uuid
        in: query
        name: manager_id
        type: string
      - description: role uuid
        in: query
        name: role_id
        type: string
      - description: filter by user.is_active
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List employees (admin)
      tags:
      - employees
    post:
      consumes:
      - application/json
      parameters:
      - description: user creds + HR fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.EmployeeCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create employee (admin) — creates user + employee in tx
      tags:
      - employees
  /api/v1/employees/{id}:
    delete:
      parameters:
      - description: employee uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Soft-delete employee + deactivate user (admin)
      tags:
      - employees
    get:
      parameters:
      - description: employee uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get employee by id (admin)
      tags:
      - employees
    patch:
      consumes:
      - application/json
      parameters:
      - description: employee uuid
        in: path
        name: id
        required: true
        type: string
      - description: fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.EmployeeUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update employee (admin)
      tags:
      - employees
  /api/v1/employees/{id}/avatar:
    patch:
      consumes:
      - multipart/form-data
      parameters:
      - description: employee uuid
        in: path
        name: id
        required: true
        type: string
      - description: image (PNG/JPEG/WEBP, max 5MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an employee's avatar (admin)
      tags:
      - employees
...
  /api/v1/employees/{id}/leave-quota:
    patch:
      consumes:
      - application/json
      parameters:
      - description: employee uuid
        in: path
        name: id
        required: true
        type: string
      - description: quotas
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.LeaveQuotaUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update employee leave quota (admin)
      tags:
      - employees
...
  /api/v1/employees/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current employee HR profile
      tags:
      - employees
    patch:
      consumes:
      - application/json
      parameters:
      - description: allowed fields only
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_exnodes_hrm-api_internal_dto.EmployeeSelfUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update current employee HR profile (restricted whitelist)
      tags:
      - employees
  /api/v1/employees/me/avatar:
    patch:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image (PNG/JPEG/WEBP, max 5MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Replace current employee avatar
      tags:
      - employees
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"